fixup! Update useEditor
